Added javadocs to everything and added run function to each manager and call in main, also fleshed out APIs.
Added and adjusted more functions in the room api... moved garbage collection to the memory api and call it from the manager. Call all maangers in the main so we can more clearly see the flow of logic from main to manager. It go like so

set up memory
run empire and take that info
run all rooms
spawn what we need to